screenshoot

+++ src/main/java/kevindonati/M5_W3_D5/services/EventoService.java
@@ -7,6 +7,7 @@ import kevindonati.M5_W3_D5.exceptions.NotFoundException;
 import kevindonati.M5_W3_D5.exceptions.UnauthorizedException;
 import kevindonati.M5_W3_D5.payloads.EventoDTO;
 import kevindonati.M5_W3_D5.repositories.EventoRepository;
+import kevindonati.M5_W3_D5.repositories.PrenotazioneRepository;
 import kevindonati.M5_W3_D5.repositories.UtenteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -25,10 +26,16 @@ public class EventoService {
     private EventoRepository eventoRepository;
     @Autowired
     private UtenteService utenteService;
+    @Autowired
+    private PrenotazioneRepository prenotazioneRepository;
 
     public Evento save(EventoDTO payload, Utente organizzatore) {
         if (payload.dataEvento().isBefore(LocalDate.now())) {
             throw new BadRequestException("La data non può essere nel passato");
+        }
+
+        if (eventoRepository.existsByTitoloAndDataEventoAndLuogo(payload.titolo(), payload.dataEvento(), payload.luogo())) {
+            throw new BadRequestException("Esiste già un evento con questo titolo, questa data e questo luogo");
         }
         Evento nuovoEvento = new Evento(payload.titolo(), payload.descrizione(), payload.dataEvento(), payload.luogo(), payload.postiDisponibili(), organizzatore);
         return eventoRepository.save(nuovoEvento);
@@ -70,6 +77,9 @@ public class EventoService {
         Evento eventoTrovato = this.findById(id);
         if (!eventoTrovato.getOrganizzatore().getId().equals(utenteAutenticato.getId())) {
             throw new UnauthorizedException("Non puoi eliminare un evento creato da un altro organizzatore");
+        }
+        if (prenotazioneRepository.existsByEvento(eventoTrovato)) {
+            throw new BadRequestException("Non puoi eliminare un evento che ha delle prenotazioni attive");
         }
         eventoRepository.delete(eventoTrovato);
     }
